Change Tag::line field from an int to an unsigned long

+++ src/ctags.h
@@ -17,7 +17,7 @@ typedef struct {
     char *pattern;
     char *member;
     char *typeref;
-    int line;
+    unsigned long line;
     char kind;
     bool local;
 } Tag;
